ServiceNLP/app/db/database.py
import os
from dotenv import load_dotenv
from snowflake.connector import connect 

load_dotenv()  # Charge les variables d'environnement depuis .env

def get_db_connection():
    # Vérifiez que toutes les variables d'environnement nécessaires sont définies
    required_env_vars = [
        "SNOWFLAKE_USER",
        "SNOWFLAKE_PASSWORD",
        "SNOWFLAKE_ACCOUNT",
        "SNOWFLAKE_WAREHOUSE",
        "SNOWFLAKE_DATABASE",
        "SNOWFLAKE_SCHEMA"
    ]
    
    missing_vars = [var for var in required_env_vars if not os.getenv(var)]
    if missing_vars:
        raise EnvironmentError(f"Les variables d'environnement suivantes sont manquantes : {', '.join(missing_vars)}")
    
    # Connexion à Snowflake
    connection = connect(
        user=os.getenv("SNOWFLAKE_USER"),
        password=os.getenv("SNOWFLAKE_PASSWORD"),
        account=os.getenv("SNOWFLAKE_ACCOUNT"),
        warehouse=os.getenv("SNOWFLAKE_WAREHOUSE"),
        database=os.getenv("SNOWFLAKE_DATABASE"),
        schema=os.getenv("SNOWFLAKE_SCHEMA"),
    )
    return connection
